+++ src/target/BaseTarget.ts
@@ -6,6 +6,7 @@ import { CompilerWrittenFile } from "../compiler/CompilerWrittenFile";
 import { Config } from "../config/Config";
 import * as path from "path";
 import * as fs from "fs";
+import { SourceFile } from "../compiler/SourceFile";
 
 declare type TypescriptExport = {
     name: string,
@@ -43,7 +44,7 @@ export abstract class BaseTarget<C extends Config = Config> implements Partial<T
     constructor(
         protected project: Project<C>,
         protected typeChecker: ts.TypeChecker,
-        protected sourceFile: ts.SourceFile
+        protected sourceFile: SourceFile
     ) { }
 
     /**